[refactor] : 마이페이지 패키지 생성 및 이동
