+++ app/src/main/java/com/umc/yourweather/presentation/mypage/MyPageFragment.kt
@@ -1,4 +1,4 @@
-package com.umc.yourweather.presentation
+package com.umc.yourweather.presentation.mypage
 
 import android.content.Intent
 import android.os.Bundle
