+++ app/src/main/java/com/umc/yourweather/presentation/mypage/MyPageNicknameChange.kt
@@ -1,4 +1,4 @@
-package com.umc.yourweather.presentation
+package com.umc.yourweather.presentation.mypage
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
