+++ app/src/main/res/layout/activity_my_page_alarm.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.MyPageAlarm">
+    tools:context=".presentation.mypage.MyPageAlarm">
 
     <ImageButton
         android:id="@+id/btn_my_alarm"
