+++ app/src/main/res/layout/activity_my_page_nickname_change.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.MyPageNicknameChange">
+    tools:context=".presentation.mypage.MyPageNicknameChange">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/cl_mypage_nickname_title"
