+++ app/src/main/res/layout/activity_my_page_pw_change.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.MyPagePwChange">
+    tools:context=".presentation.mypage.MyPagePwChange">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/cl_mypage_nickname_title"
